robot_client/requirements.txt
requests>=2.25.0
opencv-python>=4.5.0
pyzbar>=0.1.8
qrcode>=7.0
Pillow>=8.0.0
numpy>=1.20.0
python-dotenv>=0.19.0
pygame>=2.0.0
websockets>=11.0.0 